backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
redis==5.0.1
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.26.0
python-multipart==0.0.9